chore(manifest): rename manifest-fragment.utils to manifest-fragment.helper (#2750)
