+++ apps/backend/src/modules/shareable-resource/manifest-fragment/manifest-fragment.helper.test.ts
@@ -1,17 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { BadRequestErrorCode } from '../../../utils/error/error.code';
-import {
-  assertHomogeneousLtsBatch,
-  buildConnectorLogoFilename,
-  findMinConnectorVersion,
-  formatConnectorVersion,
-  getConnectorDocumentTags,
-  getConnectorMetadataFromExisting,
-  getLatestTagForConnectorVersion,
-  isStrictlyGreaterConnectorVersion,
-  validateConnectorMinimumVersion,
-  validateShortDescriptionLength,
-} from './manifest-fragment.utils';
+import { ManifestFragmentHelper } from './manifest-fragment.helper';
 
 describe('formatConnectorVersion', () => {
   it.each`
@@ -21,7 +10,7 @@ describe('formatConnectorVersion', () => {
     ${'6.5.1'}            | ${'006.000005.001'}
     ${'6.5.1-lts.2'}      | ${'006.000005.001.LTS.002'}
   `('formats "$input" as "$expected"', ({ input, expected }) => {
-    expect(formatConnectorVersion(input)).toBe(expected);
+    expect(ManifestFragmentHelper.formatConnectorVersion(input)).toBe(expected);
   });
 
   it.each`
@@ -33,7 +22,7 @@ describe('formatConnectorVersion', () => {
     ${'LTS.1.2.3'}
     ${'7.20260703.0'}
   `('throws when version format is invalid: "$input"', ({ input }) => {
-    expect(() => formatConnectorVersion(input)).toThrow(
+    expect(() => ManifestFragmentHelper.formatConnectorVersion(input)).toThrow(
       BadRequestErrorCode.InvalidConnectorVersionFormat
     );
   });
@@ -45,7 +34,9 @@ describe('validateConnectorMinimumVersion', () => {
     ${'7.260309.0'}
     ${'7.260309.0-lts.5'}
   `('accepts "$input"', ({ input }) => {
-    expect(() => validateConnectorMinimumVersion(input)).not.toThrow();
+    expect(() =>
+      ManifestFragmentHelper.validateConnectorMinimumVersion(input)
+    ).not.toThrow();
   });
 
   it.each`
@@ -54,9 +45,9 @@ describe('validateConnectorMinimumVersion', () => {
     ${'not-a-version'}
     ${'7.20260703.0'}
   `('throws for invalid value "$input"', ({ input }) => {
-    expect(() => validateConnectorMinimumVersion(input)).toThrow(
-      BadRequestErrorCode.InvalidConnectorVersionFormat
-    );
+    expect(() =>
+      ManifestFragmentHelper.validateConnectorMinimumVersion(input)
+    ).toThrow(BadRequestErrorCode.InvalidConnectorVersionFormat);
   });
 });
 
@@ -70,7 +61,9 @@ describe('validateShortDescriptionLength', () => {
     'accepts a short_description of length $length ($description)',
     ({ length }) => {
       expect(() =>
-        validateShortDescriptionLength('a'.repeat(length))
+        ManifestFragmentHelper.validateShortDescriptionLength(
+          'a'.repeat(length)
+        )
       ).not.toThrow();
     }
   );
@@ -82,9 +75,11 @@ describe('validateShortDescriptionLength', () => {
   `(
     'throws for a short_description of length $length ($description)',
     ({ length }) => {
-      expect(() => validateShortDescriptionLength('a'.repeat(length))).toThrow(
-        BadRequestErrorCode.ShortDescriptionTooLong
-      );
+      expect(() =>
+        ManifestFragmentHelper.validateShortDescriptionLength(
+          'a'.repeat(length)
+        )
+      ).toThrow(BadRequestErrorCode.ShortDescriptionTooLong);
     }
   );
 });
@@ -95,7 +90,9 @@ describe('getLatestTagForConnectorVersion', () => {
     ${'007.260309.000'}         | ${'latest'}
     ${'007.260309.000.LTS.005'} | ${'latest-lts'}
   `('returns "$expected" for "$input"', ({ input, expected }) => {
-    expect(getLatestTagForConnectorVersion(input)).toBe(expected);
+    expect(ManifestFragmentHelper.getLatestTagForConnectorVersion(input)).toBe(
+      expected
+    );
   });
 });
 
@@ -111,7 +108,7 @@ describe('isStrictlyGreaterConnectorVersion', () => {
     'compares padded candidate "$candidate" with current "$current"',
     ({ candidate, current, expected }) => {
       expect(
-        isStrictlyGreaterConnectorVersion({
+        ManifestFragmentHelper.isStrictlyGreaterConnectorVersion({
           candidate,
           current,
         })
@@ -123,7 +120,7 @@ describe('isStrictlyGreaterConnectorVersion', () => {
 describe('getConnectorMetadataFromExisting', () => {
   it('returns undefined when no metadata is available', () => {
     expect(
-      getConnectorMetadataFromExisting({
+      ManifestFragmentHelper.getConnectorMetadataFromExisting({
         currentLatestConnector: undefined,
         existingBatchConnectors: [],
       })
@@ -132,7 +129,7 @@ describe('getConnectorMetadataFromExisting', () => {
 
   it('prefers current latest metadata and falls back to existing connectors', () => {
     expect(
-      getConnectorMetadataFromExisting({
+      ManifestFragmentHelper.getConnectorMetadataFromExisting({
         currentLatestConnector: {
           datasheet_url: 'https://latest/datasheet',
         },
@@ -162,7 +159,10 @@ describe('getConnectorDocumentTags', () => {
     'builds tags for shouldPromoteAsLatest=$shouldPromoteAsLatest',
     ({ shouldPromoteAsLatest, latestTag, expected }) => {
       expect(
-        getConnectorDocumentTags(shouldPromoteAsLatest, latestTag)
+        ManifestFragmentHelper.getConnectorDocumentTags(
+          shouldPromoteAsLatest,
+          latestTag
+        )
       ).toEqual(expected);
     }
   );
@@ -171,7 +171,7 @@ describe('getConnectorDocumentTags', () => {
 describe('buildConnectorLogoFilename', () => {
   it('builds a unique filename based on connector title and version', () => {
     expect(
-      buildConnectorLogoFilename({
+      ManifestFragmentHelper.buildConnectorLogoFilename({
         title: 'MISP',
         version: '7.260309.0-lts.5',
       })
@@ -189,24 +189,29 @@ describe('findMinConnectorVersion', () => {
     [['6.5.1-lts.2', '6.5.1-lts.1'], '6.5.1-lts.1'],
     [['6.5.1', '6.5.1'], '6.5.1'],
   ])('returns %s for %s', (versions, expected) => {
-    expect(findMinConnectorVersion(versions)).toBe(expected);
+    expect(ManifestFragmentHelper.findMinConnectorVersion(versions)).toBe(
+      expected
+    );
   });
 });
 
 describe('assertHomogeneousLtsBatch', () => {
   it('returns false (non-LTS) for an empty batch', () => {
-    expect(assertHomogeneousLtsBatch([])).toBe(false);
+    expect(ManifestFragmentHelper.assertHomogeneousLtsBatch([])).toBe(false);
   });
 
   it('returns false for a non-LTS batch', () => {
     expect(
-      assertHomogeneousLtsBatch([{ version: '7.0.0' }, { version: '7.1.0' }])
+      ManifestFragmentHelper.assertHomogeneousLtsBatch([
+        { version: '7.0.0' },
+        { version: '7.1.0' },
+      ])
     ).toBe(false);
   });
 
   it('returns true for an LTS batch', () => {
     expect(
-      assertHomogeneousLtsBatch([
+      ManifestFragmentHelper.assertHomogeneousLtsBatch([
         { version: '7.260309.0-lts.5' },
         { version: '7.260309.0-lts.6' },
       ])
@@ -215,7 +220,7 @@ describe('assertHomogeneousLtsBatch', () => {
 
   it('rejects a batch mixing LTS and non-LTS fragments', () => {
     expect(() =>
-      assertHomogeneousLtsBatch([
+      ManifestFragmentHelper.assertHomogeneousLtsBatch([
         { version: '7.0.0' },
         { version: '7.260309.0-lts.5' },
       ])
